requirements.txt
python-dotenv
haystack-ai
PyPDF2
